Refactor UI with view binding and theme manager
